requirements.txt
pydantic>=2.0.0
python-dateutil>=2.8.2
colorlog>=6.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
jira>=3.0.0
